Reorganize Folder Structure
File relocation, directory creation, and updates to namespaces and references are part of this change.
