+++ CineTicketHub/Controllers/MoviesController.cs
@@ -1,15 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using CineTicketHub.Models;
-using CineTicketHub.Data.ViewModels;
 using CineTicketHub.Mappers;
-using CineTicketHub.Models.Services;
-using Microsoft.IdentityModel.Tokens;
+using CineTicketHub.Models.ViewModels;
+using CineTicketHub.Services;
 
 namespace CineTicketHub.Controllers
 {
