--- CineTicketHub/Data/Services/IGenresService.cs
@@ -1,8 +1,0 @@
-using CineTicketHub.Models.Base;
-
-namespace CineTicketHub.Models.Services;
-
-public interface IGenresService : IEntityBaseRepository<Genre>
-{
-    
-}
