--- CineTicketHub/Data/Services/IMoviesService.cs
@@ -1,10 +1,0 @@
-using CineTicketHub.Data.ViewModels;
-using CineTicketHub.Models.Base;
-
-namespace CineTicketHub.Models.Services;
-
-public interface IMoviesService : IEntityBaseRepository<Movie>
-{
-    Task AddMovieAsync(MovieVM movieVM);
-    Task UpdateMovieAsync(MovieVM movieVM);
-}
