+++ CineTicketHub/Services/IGenresService.cs
@@ -1,0 +1,10 @@
+using CineTicketHub.Models;
+using CineTicketHub.Models.Entities;
+using CineTicketHub.Services.Base;
+
+namespace CineTicketHub.Services;
+
+public interface IGenresService : IEntityBaseRepository<Genre>
+{
+    
+}
