+++ CineTicketHub/Views/Genres/Edit.cshtml
@@ -1,4 +1,4 @@
-@model CineTicketHub.Models.Genre
+@model CineTicketHub.Models.Entities.Genre
 
 @{
     ViewData["Title"] = "Edit";
