+++ CineTicketHub/Views/Genres/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<CineTicketHub.Models.Genre>
+@model IEnumerable<CineTicketHub.Models.Entities.Genre>
 
 @{
     ViewData["Title"] = "Index";
