+++ CineTicketHub/Views/Movies/Create.cshtml
@@ -1,4 +1,4 @@
-@model CineTicketHub.Data.ViewModels.MovieVM
+@model CineTicketHub.Models.ViewModels.MovieVM
 
 @{
     ViewData["Title"] = "Create";
