+++ CineTicketHub/Views/Movies/Details.cshtml
@@ -1,4 +1,4 @@
-@model CineTicketHub.Models.Movie
+@model CineTicketHub.Models.Entities.Movie
 
 @{
     ViewData["Title"] = "Details";
