+++ CineTicketHub/Views/Movies/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<CineTicketHub.Models.Movie>
+@model IEnumerable<CineTicketHub.Models.Entities.Movie>
 
 @{
     ViewData["Title"] = "Index";
