+++ CineTicketHub/Views/Reservations/Create.cshtml
@@ -1,4 +1,4 @@
-@model CineTicketHub.Models.Reservation
+@model CineTicketHub.Models.Entities.Reservation
 
 @{
     ViewData["Title"] = "Create";
