+++ CineTicketHub/Views/Reservations/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<CineTicketHub.Models.Reservation>
+@model IEnumerable<CineTicketHub.Models.Entities.Reservation>
 
 @{
     ViewData["Title"] = "Index";
