+++ CineTicketHub/Views/Rooms/Details.cshtml
@@ -1,4 +1,4 @@
-@model CineTicketHub.Models.Room
+@model CineTicketHub.Models.Entities.Room
 
 @{
     ViewData["Title"] = "Details";
