+++ CineTicketHub/Views/Rooms/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<CineTicketHub.Models.Room>
+@model IEnumerable<CineTicketHub.Models.Entities.Room>
 
 @{
     ViewData["Title"] = "Index";
