+++ CineTicketHub/Views/Screenings/Delete.cshtml
@@ -1,4 +1,4 @@
-@model CineTicketHub.Models.Screening
+@model CineTicketHub.Models.Entities.Screening
 
 @{
     ViewData["Title"] = "Delete";
