+++ CineTicketHub/Views/Screenings/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<CineTicketHub.Models.Screening>
+@model IEnumerable<CineTicketHub.Models.Entities.Screening>
 
 @{
     ViewData["Title"] = "Index";
