+++ CineTicketHub/Views/Shared/Error.cshtml
@@ -1,4 +1,4 @@
-﻿@model ErrorViewModel
+﻿@model CineTicketHub.Models.Entities.ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
